pdf_generator/requirements.txt
Django==3.2.10
djangorestframework==3.13.1
pdfkit==1.0.0
psycopg2-binary==2.9.2
celery==5.2.2
flower==1.0.0
gunicorn==20.1.0